+++ CHANGELOG.md
@@ -9,13 +9,13 @@ and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
 ### Added
 - [docs, main] Add Deezer Connection deprecation notice
-- [player] Stop player when audio stream errors occur
+- [player] Restart the client when an audio stream error occurs
 - [player] Stop player when another device becomes active
 
 ### Changed
 - [dither] Replaced own TPDF generation with Rodio's noise generator
 - [main] Disable noise shaping by default
-- [main, player] Replaced `fastrand` with `rand` for cleaner dependencies
+- [main, player] Replaced `fastrand` with `rand` for cleaner dependency tree
 - [player] Device enumeration has changed. Run with `-d "?"` to see available devices.
 - [remote] Disconnect from the controller when setting the player state fails